# docs/swagger.yaml
basePath: /accounts
definitions:
  admin.ChangePasswordInput:
    properties:
      new_password:
        type: string
      password:
        type: string
    required:
    - new_password
    - password
    type: object
  admin.ChangeTimeframeInput:
    properties:
      timeframe:
        type: string
    required:
    - timeframe
    type: object
  admin.CoinPair:
    properties:
      active:
        example: true
        type: boolean
      coin:
        example: FTMUSDT
        type: string
      id:
        example: 25
        type: integer
      leverage:
        example: 80
        type: integer
    type: object
...
  admin.LoginUserInput:
    properties:
      UserId:
        type: string
      email:
        type: string
      password:
        type: string
    required:
    - password
    type: object
  admin.OTPInput:
    properties:
      id:
        type: string
      token:
        type: string
    type: object
  admin.OTPResponse:
    properties:
      otp_enabled:
        type: string
    type: object
  admin.RegisterUserInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - name
    - password
    type: object
  admin.Resp:
    properties:
      email:
        type: string
      id:
        type: string
      otp_verified:
        type: boolean
    type: object
...
  controllers.LoginResp:
    properties:
      email:
        type: string
      id:
        type: string
      jwt_token:
        type: string
      otp_verified:
        type: boolean
    type: object
  controllers.OtpGenResp:
    properties:
      base32:
        type: string
      otpauth_url:
        type: string
    type: object
  controllers.OtpValidResp:
    properties:
      otp_valid:
        type: boolean
    type: object
  controllers.OtpVerifyResp:
    properties:
      email:
        type: string
      id:
        type: string
      otp_verified:
        type: boolean
      token:
        type: string
    type: object
  controllers.TimeframeResp:
    properties:
      timeframe:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is BidBox Accounts Service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BidBox Accounts Service API
  version: "1.0"
paths:
  /admin/changePassword:
    post:
      consumes:
      - application/json
      description: change password
      parameters:
      - description: change password
        in: body
        name: otp
        required: true
        schema:
          $ref: '#/definitions/admin.ChangePasswordInput'
      - description: user id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Change Password
      tags:
      - auth
  /admin/changetimeframe:
    post:
      consumes:
      - application/json
      description: change timeframe
      parameters:
      - description: change timeframe
        in: body
        name: otp
        required: true
        schema:
          $ref: '#/definitions/admin.ChangeTimeframeInput'
      - description: user id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Change Timeframe
      tags:
      - auth
  /admin/coins:
    delete:
      consumes:
      - application/json
      description: delete coin
      parameters:
      - description: coin pair
        in: query
        name: coin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete Coin
      tags:
      - coins
    get:
      consumes:
      - application/json
      description: get coins
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/admin.CoinPair'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get All Coins
      tags:
      - coins
    post:
      consumes:
      - application/json
      description: save coin
      parameters:
      - description: coin pair
        in: body
        name: coin
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.CoinPair'
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Save Coin
      tags:
      - coins
    put:
      consumes:
      - application/json
      description: update coin
      parameters:
      - description: coin pair
        in: body
        name: coin
        schema:
          $ref: '#/definitions/admin.CoinPair'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.CoinPair'
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update Coin
      tags:
      - coins
...
  /admin/login:
    post:
      consumes:
      - application/json
      description: login admin
      parameters:
      - description: user login body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/admin.LoginUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.LoginResp'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login Admin
      tags:
      - auth
  /admin/otp/enable:
    post:
      consumes:
      - application/json
      description: enable otp
      parameters:
      - description: otp enable
        in: body
        name: otp
        required: true
        schema:
          $ref: '#/definitions/admin.OTPResponse'
      - description: user id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.Resp'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Enable OTP
      tags:
      - auth
  /admin/otp/generate:
    post:
      consumes:
      - application/json
      description: generate otp
      parameters:
      - description: otp input
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/admin.OTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.OtpGenResp'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Generate OTP
      tags:
      - auth
  /admin/otp/validate:
    post:
      consumes:
      - application/json
      description: validate otp
      parameters:
      - description: otp input
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/admin.OTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.OtpValidResp'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Validate OTP
      tags:
      - auth
  /admin/otp/verify:
    post:
      consumes:
      - application/json
      description: verify otp
      parameters:
      - description: otp input
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/admin.OTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.OtpVerifyResp'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Verify OTP
      tags:
      - auth
  /admin/register:
    post:
      consumes:
      - application/json
      description: sign up admin user
      parameters:
      - description: user body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/admin.RegisterUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "502":
          description: Bad Gateway
          schema:
            type: string
      summary: Sign Up Admin User
      tags:
      - auth
  /admin/timeframe:
    post:
      consumes:
      - application/json
      description: get timeframe
      parameters:
      - description: user id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.TimeframeResp'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Timeframe
      tags:
      - auth
...
swagger: "2.0"